Test.Tasty.Golden.Manage

Description

Previously, accepting tests (by the --accept flag) was done by this module, and there was an ingredient to handle that mode.

Now it's done as part of a normal test run. When the `--accept` flag is given, it makes golden tests to update the files whenever there is a mismatch. So you no longer need this module. It remains only for backwards compatibility.

Command line helpers

defaultMain :: TestTree -> IO ()

This exists only for backwards compatibility. Use defaultMain instead.

The ingredient

acceptingTests :: Ingredient

This exists only for backwards compatibility. You don't need to include this anymore.

newtype AcceptTests

This option, when set to True, specifies that we should run in the «accept tests» mode

Constructors

AcceptTests Bool